modify onSelectedStar function and delete checked planet img
