+++ frontend/src/javascripts/components/atoms/planet.js
@@ -1,8 +1,24 @@
 import React from 'react'
 import { PlanetImgs } from '../../constants/images'
 
-const Planet = ({ className, planetType, onClick }) => (
-  <img className={className} src={PlanetImgs[planetType]} onClick={onClick} />
+const Planet = ({
+  className,
+  planetType,
+  onClick,
+  onMouseOver,
+  onMouseOut,
+}) => (
+  <div
+    className={className}
+    onClick={onClick}
+    onMouseOver={onMouseOver}
+    onMouseOut={onMouseOut}
+  >
+    <img className="planet" src={PlanetImgs[planetType]} />
+    <div className="mark-container">
+      <div className="check-mark" />
+    </div>
+  </div>
 )
 
 export default Planet
